+++ java-consumer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "kafka-node",
+  "name": "countries-consumer",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -8,9 +8,10 @@
     "start": "node src/index.js"
   },
   "keywords": [],
-  "author": "Luiz Henrique Lobo",
+  "author": "Luiz Henrique",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.0.3",
     "kafkajs": "^2.2.4"
   }
 }
