+++ javascript-consumer/.npmrc
