+++ javascript-consumer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "countries-consumer",
+  "name": "animals-consumer",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -8,9 +8,10 @@
     "start": "node src/index.js"
   },
   "keywords": [],
-  "author": "Luiz Henrique",
+  "author": "Luiz Henrique Lobo",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.0.3",
     "kafkajs": "^2.2.4"
   }
 }
