--- src/data.js
@@ -1,65 +1,0 @@
-export const ANIMALS = [
-  'Elefante',
-  'Girafa',
-  'Leão',
-  'Tigre',
-  'Rinoceronte',
-  'Zebra',
-  'Hipopótamo',
-  'Crocodilo',
-  'Panda',
-  'Canguru',
-  'Orangotango',
-  'Chimpanzé',
-  'Gorila',
-  'Urso Polar',
-  'Lobo',
-  'Tubarão',
-  'Golfinho',
-  'Baleia',
-  'Cavalo',
-  'Gato',
-  'Cachorro',
-  'Coelho',
-  'Rato',
-  'Esquilo',
-  'Castor',
-  'Lontra',
-  'Tartaruga',
-  'Serpente',
-  'Arara',
-  'Papagaio'
-];
-
-export const COUNTRIES = [
-  'Afeganistão',
-  'Alemanha',
-  'Argentina',
-  'Austrália',
-  'Brasil',
-  'Canadá',
-  'China',
-  'Dinamarca',
-  'Egito',
-  'Estados Unidos',
-  'França',
-  'Grécia',
-  'Índia',
-  'Itália',
-  'Japão',
-  'México',
-  'Nigéria',
-  'Noruega',
-  'Países Baixos',
-  'Peru',
-  'Polônia',
-  'Portugal',
-  'Reino Unido',
-  'Rússia',
-  'Singapura',
-  'Suécia',
-  'Suíça',
-  'Turquia',
-  'Ucrânia',
-  'Uruguai',
-];
